requirements.txt
streamlit>=1.28.0
openai>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
